Add Bouncer
Recently, I made my first game in Game Maker Studio 2 called "Bouncer", and I released it for Windows on Itch.io (LINK: https://billycrawdad.itch.io/bouncer), it's a simple game inspired by Pong and Breakout where you keep a ball from falling at the bottom of the screen for as long as possible, and breakable bricks occasionally appear on-screen. So I made a Thumby port of the game from the ground up, and I'm hoping to add it to the Thumby Arcade library of games.