Update wilderness.py
added checks to confirm that monsters have unique names. updated how roaming monster's stats are generated for fighting. When you tame a monster it's stats are the base stats before the game simulates training it. a campfire and merchant can now be found in rooms while exploring. changed chances for taming monsters.